декомпозиция Home в CitiesBlock.tsx
